Moved most photo code to model.
